+++ test-library/build.gradle
@@ -48,6 +48,9 @@ dependencies {
     implementation "com.google.dagger:hilt-android-testing:$hilt_version"
     kapt "com.google.dagger:hilt-android-compiler:$hilt_version"
 
+    // joda datetime
+    implementation "net.danlew:android.joda:$joda"
+
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation 'androidx.appcompat:appcompat:1.4.2'
     implementation 'com.google.android.material:material:1.6.1'
