+++ test-library/src/main/java/com/pocs/test_library/rule/JodaRule.kt
@@ -1,0 +1,24 @@
+package com.pocs.test_library.rule
+
+import org.joda.time.DateTimeZone
+import org.joda.time.tz.Provider
+import org.joda.time.tz.UTCProvider
+import org.junit.rules.TestRule
+import org.junit.runner.Description
+import org.junit.runners.model.Statement
+
+class JodaRule @JvmOverloads constructor(private val provider: Provider = UTCProvider()) : TestRule {
+
+    override fun apply(base: Statement, description: Description): Statement {
+
+        return object : Statement() {
+
+            @Throws(Throwable::class)
+            override fun evaluate() {
+                DateTimeZone.setProvider(provider)
+                base.evaluate()
+            }
+        }
+    }
+
+}
